# docs/swagger.yaml
basePath: /
definitions:
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.UserRegis:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk MongoDB JWT User
  title: Golang JWT Authentication API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user dengan email dan password
      parameters:
      - description: User Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses: {}
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan user baru dengan email dan password
      parameters:
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRegis'
      produces:
      - application/json
      responses: {}
      summary: Register user
      tags:
      - Auth
  /users:
    get:
      description: Mendapatkan semua data user
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
  /users/{id}:
    delete:
      description: Menghapus data user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Memperbarui data user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRegis'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"